ML/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Machine learning models
xgboost>=1.5.0
lightgbm>=3.3.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# API and serving
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Configuration and utilities
python-dotenv>=0.19.0
setuptools>=65.0.0

# For interacting with database
prisma==0.9.1
asyncpg==0.27.0
pymongo==4.4.1

# For monitoring and logging
mlflow==2.5.0

# For testing
pytest==7.4.0 